Added nchrs.xyz
Hey Devine!

I hope it is alright that I created a new pull request instead of trying to re-open the previous one with the incomplete description.
Consulting the README I was missing the description of where to find the webring icon on my website, correct?

The icon for the webring can be found in the footer at the bottom of the page. 
I'll try to add a screenshot in a following comment (can't seem to attach images to the description directly).

Thanks!