Updating index.html, adding gr0k.net
Adding gr0k.net, which has the xxiivv logo + link on the footer of every page, links to random and next xxiivv webring sites on the links page, and a link to the xxiivv webring on the about page.

Long time follower/lurker, glad to hopefully be a part of this initiative. I miss the old days.

Thanks for all your hard work!