adding freakylamps
webring icon is at the bottom of my site. I've been rejected before from the webring, but i am asking to be reconsidered. ive restructured my site since i last asked to be included.  my site is all hand done and even though it does have a few links to external sites, all of my site is coded by me with just html and css and some javascript. the main portfolio parts are the /productions and /lamps and /art