Add kxvin.com
I'd like to add my little art and project website to the webring.

The webring icon is located on the main/home page, near the bottom and next to the other svg icons. There is also text fallback for the svg icons if the browser cannot render <svg> elements.

Thank you for your consideration!