Add my website to the ring
The webring link is in the page header. It is my personal webpage, blog, and wiki-like system heavily inspired by many of the other pages in the ring.

Please note that I am in violation of the 10-page rule (when blog entries are ignored); however, I have noticed that many pages in the ring are, so I assumed this rule was not strictly enforced. Please feel free to close my request or leave it open; I will comment on it or reopen it when I think I have reached the required ten pages.