Add bbo.do
I would love to add my personal homepage, [bbo.do](https://bbo.do), to the webring. I have added the icon in a link in the global navbar, present on all subpages. It should render on the top right, next to the github icon, and the button to change the theme.