add ribo.zone
Hello! I love so many of the sites on this webring and it'd be an honor to add my own: https://ribo.zone

The link I added to `index.html` is https://ribo.zone/neighborhood , my page dedicated to exploring handcrafted websites and where I put webring links. There's some javascript to shuffle the entries on the page, but if you have it disabled, the XXIIVV webring should be in the top row.