Add stage7.net to index
Hi Devine,

I would like to add my site https://stage7.net/ to the webring. The icon can be seen in the index page, next to my 88x31, near the end. Due to the nature of my site I have used a raster version with a limited palette. Hope it is not an issue.

Thank you and best regards,
stage7 / @stage7@owo.cafe