Update domain for wolf.md to wol.fm
In an act of impulsive vanity, I purchased a domain name that is
three letters shorter and less suggestive of having medical
credentials. Update the link and rss feed for wolfmd.me to wol.fm.

I intend on maintaining both domains for the foreseeable future
but I feel obligated to make use of the cooler (it is cooler,
right?) domain whenever possible

An update to the button will be made in conjunction to this PR